Changed up the reducers file name to todos.js. I was close when I did it originally, but it was giving me errors. The Docs page has something similar to what I had, must've missed a comma/semicolon.
