+++ todo/src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import rootReducer from './reducers';
+import rootReducer from './reducers/todos';
 
 //Store => Gloabalized State
